Improved Robustness / Versatility of config-parser
Fixed off-by-one error when displaying the column which caused
a parser error.

Made it possible to just have a simple line like "key value" in the
configuration file.

Made it possible to have lines like "key# comment" or
"key; comment" in the configuration file.

Made it possible to have a line with just a "key\n" in the
configuration file.