Added 'errstr()' function and related bash script
The function can be used to retrieve a simple human-readable error
string for a given error code. This is useful when printing error
messages and not having the urge to fiddle with 'strerror()'.

An additional auxiliary script is provided to automatically generate
the mapping table which is used for the error string lookup.